+++ Src/Public/Invoke-AsBuiltReport.Microsoft.AD.ps1
@@ -67,7 +67,6 @@ function Invoke-AsBuiltReport.Microsoft.AD {
 
     Get-RequiredModule -Name PSPKI -Version '3.7.2'
 
-
     # Import Report Configuration
     $script:Report = $ReportConfig.Report
     $script:InfoLevel = $ReportConfig.InfoLevel
@@ -75,6 +74,34 @@ function Invoke-AsBuiltReport.Microsoft.AD {
 
     # Used to set values to TitleCase where required
     $script:TextInfo = (Get-Culture).TextInfo
+
+    # Check the install status of Graphviz
+    if ($Options.EnableDiagrams) {
+        $GraphVizPath = (
+            'C:\Program Files\NuGet\Packages\Graphviz*\dot.exe',
+            'C:\program files*\GraphViz*\bin\dot.exe'
+        )
+
+        try {
+            # Use Resolve-Path to test all passed paths
+            # Select only items with 'dot' BaseName and use first one
+            $graphViz = Resolve-Path -path $GraphVizPath -ErrorAction SilentlyContinue | Get-Item | Where-Object BaseName -eq 'dot' | Select-Object -First 1
+
+            if ( $null -eq $graphViz ) {
+                $GraphvizPathString = $GraphVizPath -Join " or "
+                Write-PScriboMessage -IsWarning "Could not find GraphViz installed on this system. Please install latest Graphviz binary from: https://graphviz.org/download/#windows"
+                Write-PScriboMessage -IsWarning "No GraphViz binary found, disabling the creation of diagrams."
+
+                $GraphvizInstallStatus = $false
+            } else {
+                Write-PScriboMessage "GraphViz binary found, enabling the creation of diagrams."
+                $GraphvizInstallStatus = $true
+            }
+
+        } catch {
+            Write-PscriboMessage -IsWarning "$($_.Exception.Message) (Graphviz Install Validation)"
+        }
+    } else {$GraphvizInstallStatus = $false}
 
     #---------------------------------------------------------------------------------------------#
     #                                 Connection Section                                          #